refactor: isolate wandb field policies
